Split UDF transpile tests into their own file, gate to non-Connect
Move the five UDF transpilation test methods out of BaseUDFTestsMixin
in test_udf.py into a new test_udf_transpile_unit.py so they are no
longer inherited by the Spark Connect parity test class. Transpilation
is currently only supported in regular (non-Connect) Spark.

Also gate test_udf_transpile_hypothesis.py on is_remote_only() so the
property-based suite skips cleanly on a pyspark-client-only install,
and register the new module in dev/sparktestsupport/modules.py.
