+++ python/pyspark/sql/tests/test_udf_transpile_unit.py
@@ -1,0 +1,245 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+"""
+Unit tests for UDF transpilation.
+
+These were previously interleaved with the broader UDF mixin in
+``test_udf.py``. They are split out because UDF transpilation is currently
+only supported in regular (non-Connect) Spark, so they should not be
+inherited into the Spark Connect parity test class. The companion
+property-based suite lives in ``test_udf_transpile_hypothesis.py``.
+"""
+
+import unittest
+
+from pyspark.sql import Row
+from pyspark.sql.types import (
+    BooleanType,
+    DoubleType,
+    LongType,
+)
+from pyspark.sql.udf import UserDefinedFunction
+from pyspark.testing.sqlutils import ReusedSQLTestCase
+from pyspark.util import is_remote_only
+
+
+@unittest.skipIf(
+    is_remote_only(),
+    "UDF transpilation is only supported in regular (non-Connect) Spark.",
+)
+class UDFTranspileUnitTests(ReusedSQLTestCase):
+    def test_udf_transpile_basic(self):
+        # Test callable object
+        class PlusFour:
+            def __call__(self, col):
+                return col + 4
+
+        with self.sql_conf({"spark.sql.experimental.optimizer.transpilePyUDFS": True}):
+            # Make sure we can transpile the object
+            call = PlusFour()
+            pudf = UserDefinedFunction(call, LongType())
+            self.assertTrue(pudf.transpiled)
+            # Now make sure we can run the transpiled UDF*
+            input_df = self.spark.createDataFrame([Row(a=1)])
+            transformed_df = input_df.select(pudf("a"))
+            [row] = transformed_df.collect()
+            transformed_df.explain()
+            self.assertEqual(row[0], 5)
+
+
+        with self.sql_conf({"spark.sql.experimental.optimizer.transpilePyUDFS": False}):
+            call = PlusFour()
+            pudf = UserDefinedFunction(call, LongType())
+            self.assertEqual([], pudf.transpiled)
+            # Now make sure we can run the UDF
+            input_df = self.spark.createDataFrame([Row(a=1)])
+            transformed_df = input_df.select(pudf("a"))
+            [row] = transformed_df.collect()
+            self.assertEqual(row[0], 5)
+
+    def test_udf_transpile_with_nones(self):
+        # Test callable object
+        class PlusFour:
+            def __call__(self, col):
+                if col is None:
+                    return col + 4
+
+        with self.sql_conf({"spark.sql.experimental.optimizer.transpilePyUDFS": True}):
+            # Make sure we can transpile the object
+            call = PlusFour()
+            pudf = UserDefinedFunction(call, LongType())
+            self.assertTrue(pudf.transpiled)
+            # Now make sure we can run the transpiled UDF*
+            input_df = self.spark.createDataFrame([Row(a=1)])
+            transformed_df = input_df.select(pudf("a"))
+            [row] = transformed_df.collect()
+            self.assertEqual(row[0], 5)
+            self.assertEqual(
+                "notfound",
+                transformed_df.explain(extended=True))
+
+
+        with self.sql_conf({"spark.sql.experimental.optimizer.transpilePyUDFS": False}):
+            call = PlusFour()
+            pudf = UserDefinedFunction(call, LongType())
+            self.assertEqual([], pudf.transpiled)
+            # Now make sure we can run the UDF
+            input_df = self.spark.createDataFrame([Row(a=1)])
+            transformed_df = input_df.select(pudf("a"))
+            [row] = transformed_df.collect()
+            self.assertEqual(row[0], 5)
+
+
+    def test_udf_not_transpilable(self):
+        class UnsupportedEx:
+            def __call__(self, col):
+                if col is not None:
+                    return col in "4"
+        with self.sql_conf({"spark.sql.experimental.optimizer.transpilePyUDFS": True}):
+            call = UnsupportedEx()
+            pudf = UserDefinedFunction(call, BooleanType())
+            self.assertEqual([], pudf.transpiled)
+
+    def test_udf_transpile_requires_ansi(self):
+        # Transpilation targets ANSI semantics. With ANSI off the transpiler
+        # must skip rewriting (and warn the user) so we don't silently
+        # diverge from the Python interpretation; with ANSI on it should
+        # produce a Catalyst expression.
+        import warnings
+
+        def plus_four(x):
+            if x is not None:
+                return x + 4
+
+        with self.sql_conf({
+            "spark.sql.experimental.optimizer.transpilePyUDFS": True,
+            "spark.sql.ansi.enabled": False,
+        }):
+            with warnings.catch_warnings(record=True) as caught:
+                warnings.simplefilter("always")
+                pudf = UserDefinedFunction(plus_four, LongType())
+            self.assertEqual([], pudf.transpiled)
+            ansi_warnings = [w for w in caught if "ANSI mode" in str(w.message)]
+            self.assertTrue(
+                ansi_warnings,
+                "expected an 'ANSI mode' warning when transpilation is "
+                "requested but ANSI is disabled",
+            )
+
+        with self.sql_conf({
+            "spark.sql.experimental.optimizer.transpilePyUDFS": True,
+            "spark.sql.ansi.enabled": True,
+        }):
+            pudf = UserDefinedFunction(plus_four, LongType())
+            self.assertTrue(
+                pudf.transpiled,
+                "expected transpilation to produce a Catalyst expression "
+                "when both transpilePyUDFS and ANSI mode are enabled",
+            )
+
+    def test_udf_transpile_falls_back_for_unsupported_patterns(self):
+        # The transpiler intentionally only handles a small subset of
+        # Python AST today. Everything outside that subset must
+        # gracefully fall back to interpreted Python (with an empty
+        # `transpiled` list and a UserWarning) rather than break the
+        # UDF -- the "don't break peoples Spark code" promise. This test
+        # walks the most common unsupported shapes, registers each as a
+        # UDF with transpilation on, and asserts (a) construction does
+        # not raise, (b) `transpiled == []`, (c) the UDF still produces
+        # the correct interpreted result.
+
+        def divide_by_two(x):  # `/` -- ast.Div, not handled.
+            if x is not None:
+                return x / 2
+
+        def floor_divide_by_two(x):  # `//` -- ast.FloorDiv, not handled.
+            if x is not None:
+                return x // 2
+
+        def bit_and_one(x):  # `&` -- ast.BitAnd, not handled.
+            if x is not None:
+                return x & 1
+
+        def bit_or_one(x):  # `|` -- ast.BitOr, not handled.
+            if x is not None:
+                return x | 1
+
+        def left_shift(x):  # `<<` -- ast.LShift, not handled.
+            if x is not None:
+                return x << 1
+
+        def less_than_zero(x):  # ast.Lt, not handled (only Is/IsNot supported).
+            if x is not None:
+                return x < 0
+
+        def multi_statement(x):  # > 1 top-level statement, not handled.
+            y = 1
+            return x + y if x is not None else 0
+
+        def closure_capture(x):
+            offset = 7  # noqa: F841 -- intentional free variable
+            if x is not None:
+                return x + offset
+
+        cases = [
+            ("divide_by_two", divide_by_two, DoubleType(), Row(a=4.0), 2.0),
+            ("floor_divide_by_two", floor_divide_by_two, LongType(), Row(a=5), 2),
+            ("bit_and_one", bit_and_one, LongType(), Row(a=5), 1),
+            ("bit_or_one", bit_or_one, LongType(), Row(a=4), 5),
+            ("left_shift", left_shift, LongType(), Row(a=3), 6),
+            ("less_than_zero", less_than_zero, BooleanType(), Row(a=-1), True),
+            ("multi_statement", multi_statement, LongType(), Row(a=5), 6),
+            ("closure_capture", closure_capture, LongType(), Row(a=10), 17),
+        ]
+
+        with self.sql_conf({
+            "spark.sql.experimental.optimizer.transpilePyUDFS": True,
+            "spark.sql.ansi.enabled": True,
+        }):
+            for label, func, return_type, row, expected in cases:
+                with self.subTest(case=label):
+                    import warnings as _warnings
+                    with _warnings.catch_warnings(record=True) as caught:
+                        _warnings.simplefilter("always")
+                        pudf = UserDefinedFunction(func, return_type)
+                    self.assertEqual(
+                        [], pudf.transpiled,
+                        f"{label}: transpiler should not produce a Catalyst "
+                        "expression for this AST shape",
+                    )
+                    fallback = [
+                        w for w in caught
+                        if "Unable to transpile" in str(w.message)
+                        or "Errors encountered" in str(w.message)
+                        or "Exception transpiling" in str(w.message)
+                    ]
+                    self.assertTrue(
+                        fallback,
+                        f"{label}: expected a fallback warning when the "
+                        "transpiler can't lower the function",
+                    )
+                    df = self.spark.createDataFrame([row])
+                    [result] = df.select(pudf("a")).collect()
+                    self.assertEqual(
+                        result[0], expected,
+                        f"{label}: interpreted UDF result diverged from expected",
+                    )
+
+
+if __name__ == "__main__":
+    from pyspark.testing import main
+    main()
